Moved test to dedicated file
