+++ router/src/lib.rs
@@ -2,3 +2,4 @@ mod common_types;
 mod management;
 mod router;
 mod upgrade;
+mod tests;
